refactor(route/hko): colocate helper as utils and relocate test
Move the HKO editorial helpers from the top-level lib/ directory into
lib/routes/hko/utils.ts to follow the established namespace convention
of co-locating shared route helpers (e.g. 3dmgame/utils.ts,
163/utils.ts). Relocate the unit test out of lib/routes/ since the dev
registry dynamically imports every .ts file under that tree, which
would otherwise break route loading at runtime.

Changes by file:
- lib/routes/hko/utils.ts: renamed from lib/hko-editorial.ts
- lib/routes/hko/editorial.ts: import helpers from ./utils
- lib/hko-editorial.test.ts: import helpers from ./routes/hko/utils
